src/agentlayer/pyproject.toml
[project]
name = "agentlayer"
version = "0.1.0"
description = "Shared infrastructure for Substrate agents"
authors = [{ name = "Mike Waters", email = "mike@mikewaters.net" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.11.0",
    "loguru>=0.7.0",
    "sqlalchemy>=2.0",
    "llama-index-core>=0.12.0",
    "mlx-embeddings>=0.0.5",
]

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.hatch.build.targets.wheel]
packages = [
  "agentlayer",
]
